Android App - Java
A game created in Java using a custom SurfaceView. This game is a collaborative project between 3 other programmers using SVN to commit code. My role in the team was to develop a custom SurfaceView class to draw a canvas to the screen to have full control of what is being drawn to the screen. I also implemented the game loop with multi-threading for the draw and update methods within the game, on top of that I have implemented a class for animating sprites by extracting cells within a sprite sheet.